Add repeat id handling and consolidate message prep in utils
