+++ devtools/client/webconsole/new-console-output/actions/messages.js
@@ -3,13 +3,16 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
 "use strict";
+
+const {
+  prepareMessage
+} = require("devtools/client/webconsole/new-console-output/utils/messages");
 
 const {
   MESSAGE_ADD,
   MESSAGES_CLEAR,
-  LEVELS,
-  SEVERITY_CLASS_FRAGMENTS
 } = require("../constants");
 
 function messageAdd(packet) {
@@ -26,33 +29,5 @@ function messagesClear() {
   };
 }
 
-function prepareMessage(packet) {
-  let allowRepeating;
-  let category;
-  let data;
-  let messageType;
-  let severity;
-
-  switch (packet.type) {
-    case "consoleAPICall":
-      allowRepeating = true;
-      category = "console";
-      data = Object.assign({}, packet.message);
-      messageType = "ConsoleApiCall";
-      severity = SEVERITY_CLASS_FRAGMENTS[LEVELS[packet.message.level]];
-      break;
-  }
-
-  return {
-    allowRepeating,
-    category,
-    data,
-    messageType,
-    severity
-  };
-}
-
 exports.messageAdd = messageAdd;
 exports.messagesClear = messagesClear;
-// Export for use in testing.
-exports.prepareMessage = prepareMessage;
