+++ devtools/client/webconsole/new-console-output/moz.build
@@ -7,6 +7,7 @@ DIRS += [
     'actions',
     'components',
     'reducers',
+    'utils',
 ]
 
 DevToolsModules(
