+++ devtools/client/webconsole/new-console-output/test/components/test_console-api-call.html
@@ -13,7 +13,7 @@
 
 <script type="text/javascript;version=1.8">
 window.onload = Task.async(function* () {
-  const { prepareMessage } = require("devtools/client/webconsole/new-console-output/actions/messages");
+  const { prepareMessage } = require("devtools/client/webconsole/new-console-output/utils/messages");
   const { ConsoleApiCall } = require("devtools/client/webconsole/new-console-output/components/message-types/console-api-call");
 
   const packet = yield getPacket("console.log('foobar', 'test')", "consoleAPICall");
@@ -23,7 +23,7 @@ window.onload = Task.async(function* () {
   const queryPath = "span span.message-flex-body span.message-body.devtools-monospace";
   const messageBody = rendered.querySelectorAll(queryPath);
   const consoleStringNodes = messageBody[0].querySelectorAll("span.console-string");
-  is(2, consoleStringNodes.length, "ConsoleApiCall outputs expected HTML structure");
+  is(consoleStringNodes.length, 2, "ConsoleApiCall outputs expected HTML structure");
   is(messageBody[0].textContent, "foobar test", "ConsoleApiCall outputs expected text");
 
   SimpleTest.finish()
