+++ devtools/client/webconsole/new-console-output/test/components/test_message-container.html
@@ -15,7 +15,7 @@
 window.onload = Task.async(function* () {
   const React = require("devtools/client/shared/vendor/react");
   const TestUtils = React.addons.TestUtils;
-  const { prepareMessage } = require("devtools/client/webconsole/new-console-output/actions/messages");
+  const { prepareMessage } = require("devtools/client/webconsole/new-console-output/utils/messages");
 
   const { MessageContainer } = require("devtools/client/webconsole/new-console-output/components/message-container");
   const { ConsoleApiCall } = require("devtools/client/webconsole/new-console-output/components/message-types/console-api-call");
